Lecture 4/requirements.txt
langchain==0.3.13
langchain-google-genai==2.0.6
langchain-community==0.3.13
sentence-transformers==3.3.1
chromadb==0.5.23
python-dotenv==1.0.1
redis==5.2.1
PyPDF2==3.0.1
google-generativeai==0.8.3
